Restore old getAdditionTimestamp behavior
a190c607ad1b1981bd1337c0f1442aab9ee05a08 modified one of the
getAdditionTimestamp methods in DistributedStoreManager.  Instead of
returning the timestamp with the LSB forced on, it just returned the
timestamp.  The LSB never matters at us and ns precision, but
neglecting it can cause plausibly correctness issues at ms precision
(and it has before in test cases).  Tweaked the affected
getAdditionTimestamp method to set the LSB.

For #1076 (#1019)